refactor: astro adapter to strategy api
