+++ inlang/packages/paraglide/paraglide-astro/example/astro.config.mjs
@@ -1,34 +1,18 @@
 import paraglideAstro from "@inlang/paraglide-astro";
 import { defineConfig } from "astro/config";
-import sitemap from "@astrojs/sitemap";
 import svelte, { vitePreprocess } from "@astrojs/svelte";
 import node from "@astrojs/node";
 
 // https://astro.build/config
 export default defineConfig({
-	//configure this to your domain name
-	//site: "https://acme.com",
-
-	i18n: {
-		defaultLocale: "en",
-		locales: ["en", "de"],
-	},
 	integrations: [
-		sitemap({
-			i18n: {
-				defaultLocale: "en",
-				locales: {
-					en: "en-US",
-					de: "de-CH",
-				},
-			},
-		}),
 		svelte({
 			preprocess: [vitePreprocess()],
 		}),
 		paraglideAstro({
 			project: "./project.inlang",
 			outdir: "./src/paraglide",
+			strategy: ["pathname", "baseLocale"],
 		}),
 	],
 	output: "server",
