+++ inlang/packages/paraglide/paraglide-astro/example/package.json
@@ -1,30 +1,29 @@
 {
-  "name": "@inlang/paraglide-astro-example",
-  "type": "module",
-  "version": "0.1.0",
-  "private": "true",
-  "scripts": {
-    "dev": "astro dev",
-    "start": "astro dev",
-    "build": "paraglide-js compile --project ./project.inlang && astro check && astro build",
-    "preview": "astro preview",
-    "astro": "astro"
-  },
-  "dependencies": {
-    "@astrojs/node": "^9.0.1",
-    "@astrojs/sitemap": "3.2.1",
-    "@astrojs/svelte": "^7.0.3",
-    "astro": "^5.1.6",
-    "@astrojs/check": "^0.9.4",
-    "typescript": "^5.5.2",
-    "svelte": "^5.17.5"
-  },
-  "files": [
-    "*",
-    ".gitignore"
-  ],
-  "devDependencies": {
-    "@inlang/paraglide-astro": "workspace:*",
-    "@inlang/paraglide-js": "workspace:*"
-  }
+	"name": "@inlang/paraglide-astro-example",
+	"type": "module",
+	"version": "0.1.0",
+	"private": "true",
+	"scripts": {
+		"dev": "astro dev",
+		"start": "astro dev",
+		"build": "paraglide-js compile --project ./project.inlang && astro check && astro build",
+		"preview": "astro preview",
+		"astro": "astro"
+	},
+	"dependencies": {
+		"@astrojs/node": "^9.0.1",
+		"@astrojs/svelte": "^7.0.3",
+		"astro": "^5.1.6",
+		"@astrojs/check": "^0.9.4",
+		"typescript": "^5.5.2",
+		"svelte": "^5.17.5"
+	},
+	"files": [
+		"*",
+		".gitignore"
+	],
+	"devDependencies": {
+		"@inlang/paraglide-astro": "workspace:*",
+		"@inlang/paraglide-js": "workspace:*"
+	}
 }
