--- inlang/packages/paraglide/paraglide-astro/src/alias.ts
@@ -1,8 +1,0 @@
-export function alias(map: Record<string, string>) {
-	return {
-		name: "astro-plugin-paraglide-alias",
-		resolveId(id: string) {
-			return map[id];
-		},
-	};
-}
