+++ inlang/packages/paraglide/paraglide-astro/src/normalize-path.ts
